\name{evd-defunct}

\alias{ffrechet}
\alias{fgumbel}
\alias{frweibull}
\alias{fbvall}

\title{Defunct Functions}

\description{
  The fitting functions \code{ffrechet}, \code{fgumbel} and
  \code{frweibull} have been defunct since Version 1.2-0.
  The bivariate fitting function \code{fbvall} has been defunct
  since Version 1.2-3.
}

\keyword{internal}
