\name{failure}
\alias{failure}
\title{Failure Times}
\usage{data(failure)}
\description{
  Failure times.
}
\format{A vector containing 24 observations.}
\source{
  van Montfort, M. A. J. and Otten, A. (1978)
  On testing a shape parameter in the presence of a scale
  parameter.
  \emph{Math. Operations Forsch. Statist., Ser. Statistics},
  \bold{9}, 91--104.
}
  
\keyword{datasets}
