\name{fox}
\alias{fox}
\title{Maximum Annual Flood Discharges of the Fox River}
\usage{fox}
\description{
  The \code{fox} data frame has 33 rows and 2 columns.
  The columns contain maximum annual flood discharges, in units
  of 1000 cubed feet per second, from the Fox River in Wisconsin,
  USA at Berlin (upstream) and Wrightstown (downstream), for the
  years 1918 to 1950.
  The row names give the years of observation.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{berlin}{A numeric vector containing maximum annual flood
      discharges at Berlin (upstream).}
    \item{wright}{A numeric vector containing maximum annual flood
      discharges at Wrightstown (downstream).}
  }
}
\source{
  Gumbel, E. J. and Mustafi, C. K. (1967)
  Some analytical properties of bivariate extremal distributions.
  \emph{J. Amer. Statist. Assoc.}, \bold{62}, 569--588.
}
\keyword{datasets}
