\name{lisbon}
\alias{lisbon}
\title{Annual Maximum Wind Speeds at Lisbon}
\usage{data(lisbon)}
\description{
  A numeric vector containing annual maximum wind speeds, in
  kilometers per hour, from 1941 to 1970 at Lisbon, Portugal.
}
\format{A vector containing 30 observations.}
\source{
  Tiago de Oliveira, J. (1997)
  \emph{Statistical Analysis of Extremes.}
  Pendor.
}
\keyword{datasets}