\name{oldage}
\alias{oldage}
\title{Oldest Ages for Swedish Males and Females}
\usage{data(oldage)}
\description{
  The \code{oldage} data frame has 66 rows and 2 columns.
  The columns contain the oldest ages at death for men and women in
  Sweden, for the period 1905--1970.
  The row names give the years of observation.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{men}{A numeric vector containing the oldest ages at death
      for men.}
    \item{women}{A numeric vector containing the oldest ages at death
      for women.}
  }
}
\source{
  Fransen, A. and Tiago de Oliveira, J. (1984)
  Statistical choice of univariate extreme models, part II, in
  \emph{Statistical Extremes and Applications},
  J. Tiago de Oliveira ed., 373--394, D. Reidel, Dordrect.
}
\keyword{datasets}
