\name{oxford}
\alias{oxford}
\title{Annual Maximum Temperatures at Oxford}
\usage{data(oxford)}
\description{
  A numeric vector containing annual maximum temperatures, in degrees
  Fahrenheit, from 1901 to 1980 at Oxford, England.
}
\format{A vector containing 80 observations.}
\source{
  Tabony, R. C. (1983)
  Extreme value analysis in meteorology.
  \emph{The Meteorological Magazine} \bold{112}, 77--98.
}  
\keyword{datasets}