\name{portpirie}
\alias{portpirie}
\title{Annual Maximum Sea Levels at Port Pirie}
\usage{data(portpirie)}
\description{
  A numeric vector containing annual maximum sea levels, in metres,
  from 1923 to 1987 at Port Pirie, South Australia.
}
\format{A vector containing 65 observations.}
\source{
  Tawn, J. A. (1993)
  Extreme sea-levels, in \emph{Statistics in the Environment},
  243--263, eds. V. Barnett and F. Turkman, Wiley.
}
\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modeling of Extreme Values}.
  London: Springer-Verlag.
}
  
\keyword{datasets}