\name{sask}
\alias{sask}
\title{Maximum Annual Flood Discharges of the North Saskachevan River}
\usage{sask}
\description{
  A numeric vector containing maximum annual flood discharges, in
  units of 1000 cubic feet per second, of the North Saskachevan
  River at Edmonton, over a period of 47 years. Unfortunately,
  the data are ordered from largest to smallest.
}
\format{A vector containing 47 observations.}
\source{
  van Montfort, M. A. J. (1970)
  On testing that the distribution is of type I when type II
  is the alternative.
  \emph{J. Hydrology}, \bold{11}, 421--427.
}

  
\keyword{datasets}
