\name{sealevel}
\alias{sealevel}
\title{Annual Sea Level Maxima at Dover and Harwich}
\usage{data(sealevel)}
\description{
  The \code{sealevel} data frame has 39 rows and 2 columns.
  The columns contain annual sea level maxima from 1938 to 1976 at
  Dover and Harwich respectively, two sites on the coast of Britain.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{dover}{a numeric vector containing annual sea level maxima
    at Dover}
    \item{harwich}{a numeric vector containing sea annual level maxima
    at Harwich}
  }
}
\source{
  Coles, S. G. and Tawn, J. A. (1990)
  Statistics of coastal flood prevention.
  \emph{Phil. Trans. R. Soc. Lond., A} \bold{332}, 457--476.
}
\keyword{datasets}
