\name{std.errors}
\alias{std.errors}

\title{Model Standard Errors}
\description{
  Returns the standard errors from a fitted model object.
}
\usage{
std.errors(object, ...)
}
\arguments{
  \item{object}{An object for which the standard errors are desired.}
  \item{\dots}{Additional arguments.}
}
\value{
  The standard errors extracted from \code{object}.
}

\keyword{models}