\name{uccle}
\alias{uccle}
\title{Rainfall Maxima at Uccle, Belgium}
\usage{uccle}
\description{
  The \code{uccle} data frame has 35 rows and 4 columns.
  The columns contain annual rainfall maxima (in millimetres) from
  1938 to 1972 at Uccle, Belgium, over the durations of one day,
  one hour, ten minutes and one minute.
  The row names give the years of observation.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{day}{Annual daily rainfall maxima.}
    \item{hour}{Annual hourly rainfall maxima.}
    \item{tmin}{Annual rainfall maxima over ten minute durations.}
    \item{min}{Annual rainfall maxima over one minute durations.}
  }
}
\source{
  Sneyers, R. (1977)
  L'intensit\'{e} maximale des pr\'{e}cipitations en Belgique.
  \emph{Inst. Royal M\'{e}t\'{e}or. Belgique, B} \bold{86}.
}
\keyword{datasets}
