\name{venice}
\alias{venice}
\title{Largest Sea Levels in Venice}
\usage{data(venice)}
\description{
  The \code{venice} data frame has 51 rows and 10 columns.
  The jth column contains the jth largest sea levels in Venice,
  for the years 1905--1992.
  Only the largest six measurements are available for the year 1935;
  the corresponding row contains four missing values.
}
\format{A data frame with 51 rows and 10 columns.}
\source{
  Smith, R. L. (1986)
  Extreme value theory based on the \eqn{r} largest annual events.
  \emph{Journal of Hydrology}, \bold{86}, 27--43.
}
\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modeling of Extreme Values}.
  London: Springer-Verlag.
}
\keyword{datasets}
