\name{venice2}
\alias{venice2}
\title{Largest Sea Levels in Venice}
\usage{venice2}
\description{
  The \code{venice2} data frame has 125 rows and 10 columns.
  The data was kindly provided by Anthony Davison.
  The jth column contains the jth largest sea levels in Venice,
  for the years 1887--2011.
  This is a larger version of the dataset \code{venice}.
  Only the largest six measurements are available for the year 1935,
  and only the largest is available for 1922; the corresponding rows 
  contain missing values.
  The years for each set of measurements are given as row names. 
}
\format{A data frame with 125 rows and 10 columns.}
\source{
  Smith, R. L. (1986)
  Extreme value theory based on the \eqn{r} largest annual events.
  \emph{Journal of Hydrology}, \bold{86}, 27--43.
}
\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modeling of Extreme Values}.
  London: Springer-Verlag.
}
\keyword{datasets}
