\name{sealevel}
\alias{sealevel}
\title{Annual Sea Level Maxima at Dover and Harwich}
\usage{data(sealevel)}
\description{
  The \code{sealevel} data frame has 81 rows and 2 columns.
  The columns contain annual sea level maxima from 1912 to 1992 at
  Dover and Harwich respectively, two sites on the coast of Britain.
  There are 39 missing values.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{dover}{A numeric vector containing annual sea level maxima
    at Dover, including 9 missing values.}
    \item{harwich}{A numeric vector containing sea annual level maxima
    at Harwich, including 30 missing values.}
  }
}
\source{
  Coles, S. G. and Tawn, J. A. (1990)
  Statistics of coastal flood prevention.
  \emph{Phil. Trans. R. Soc. Lond., A} \bold{332}, 457--476.
}
\keyword{datasets}
